--- .prettierrc.json
@@ -1,9 +1,0 @@
-{
-  "semi": true,
-  "singleQuote": true,
-  "trailingComma": "all",
-  "useTabs": false,
-  "tabWidth": 2,
-  "printWidth": 80,
-  "arrowParens": "always"
-}
