+++ app/_container/Hello/Hello.module.scss
@@ -13,16 +13,15 @@ html {
     .title {
       font-size: 2rem;
       font-weight: 700;
-      color: hsl(var(--text));
+      color: var(--foreground);
     }
     .description {
       font-size: 1.25rem;
       font-weight: 500;
-      color: hsl(var(--text));
+      color: var(--foreground);
       &.sub {
         font-size: 1rem;
       }
     }
   }
-
 }
