+++ layout/BaseLayout/BaseLayout.module.scss
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   min-height: 100vh;
-  background-color: hsl(var(--background));
+  background-color: var(--background);
   .contents {
     position: relative;
     flex: 1 1 0;
