+++ layout/BaseLayout/Header/Header.module.scss
@@ -5,7 +5,7 @@
   z-index: 50;
   width: 100%;
   height: 3.5rem;
-  border-bottom: 1px solid hsl(var(--border) / 0.8);
+  border-bottom: 1px solid oklch(from var(--border) l c h / 0.3);
   backdrop-filter: blur(10px);
   .contents {
     position: relative;
