+++ prettier.config.js
@@ -1,0 +1,7 @@
+import preset from '@c0zycr473/prettier-config';
+
+const config = {
+  ...preset,
+};
+
+export default config;
